# Configurable Message.Key and Message.Value bytes allocation functions