--- a/batch.go
+++ b/batch.go
@@ -38,6 +38,11 @@ type Batch struct {
 	// we get an EOF we do not get the lastOffset. So there is a mismatch
 	// between when we receive it and need to use it.
 	lastOffset int64
+
+	// makeKeyBytes is used to create a bytes slice for Message.Key with given length.
+	makeKeyBytes func(length int) []byte
+	// makeValBytes is used to create a bytes slice for Message.Value with given length.
+	makeValBytes func(length int) []byte
 }
 
 // Throttle gives the throttling duration applied by the kafka server on the
@@ -199,11 +204,11 @@ func (batch *Batch) ReadMessage() (Message, error) {
 
 	offset, timestamp, headers, err = batch.readMessage(
 		func(r *bufio.Reader, size int, nbytes int) (remain int, err error) {
-			msg.Key, remain, err = readNewBytes(r, size, nbytes)
+			msg.Key, remain, err = readIntoAllocatedBytes(r, size, nbytes, batch.makeKeyBytes)
 			return
 		},
 		func(r *bufio.Reader, size int, nbytes int) (remain int, err error) {
-			msg.Value, remain, err = readNewBytes(r, size, nbytes)
+			msg.Value, remain, err = readIntoAllocatedBytes(r, size, nbytes, batch.makeValBytes)
 			return
 		},
 	)
@@ -215,11 +220,11 @@ func (batch *Batch) ReadMessage() (Message, error) {
 		}
 		offset, timestamp, headers, err = batch.readMessage(
 			func(r *bufio.Reader, size int, nbytes int) (remain int, err error) {
-				msg.Key, remain, err = readNewBytes(r, size, nbytes)
+				msg.Key, remain, err = readIntoAllocatedBytes(r, size, nbytes, batch.makeKeyBytes)
 				return
 			},
 			func(r *bufio.Reader, size int, nbytes int) (remain int, err error) {
-				msg.Value, remain, err = readNewBytes(r, size, nbytes)
+				msg.Key, remain, err = readIntoAllocatedBytes(r, size, nbytes, batch.makeValBytes)
 				return
 			},
 		)