--- a/message_test.go
+++ b/message_test.go
@@ -12,11 +12,12 @@ import (
 	"testing"
 	"time"
 
+	"github.com/stretchr/testify/require"
+
 	"github.com/segmentio/kafka-go/compress/gzip"
 	"github.com/segmentio/kafka-go/compress/lz4"
 	"github.com/segmentio/kafka-go/compress/snappy"
 	"github.com/segmentio/kafka-go/compress/zstd"
-	"github.com/stretchr/testify/require"
 )
 
 // This regression test covers reading messages using offsets that