--- a/read.go
+++ b/read.go
@@ -141,7 +141,9 @@ func readBytesWith(r *bufio.Reader, sz int, cb func(*bufio.Reader, int, int) (in
 	return cb(r, sz, n)
 }
 
-func readNewBytes(r *bufio.Reader, sz int, n int) ([]byte, int, error) {
+// readIntoAllocatedBytes reads n bytes from r into a slice provided by makeBytes, which should should provide a slice with the required length.
+// This slice is then returned, as well as the number of bytes remaining and an error, if any.
+func readIntoAllocatedBytes(r *bufio.Reader, sz int, n int, makeBytes func(length int) []byte) ([]byte, int, error) {
 	var err error
 	var b []byte
 	var shortRead bool
@@ -152,7 +154,7 @@ func readNewBytes(r *bufio.Reader, sz int, n int) ([]byte, int, error) {
 			shortRead = true
 		}
 
-		b = make([]byte, n)
+		b = makeBytes(n)
 		n, err = io.ReadFull(r, b)
 		b = b[:n]
 		sz -= n
@@ -163,6 +165,10 @@ func readNewBytes(r *bufio.Reader, sz int, n int) ([]byte, int, error) {
 	}
 
 	return b, sz, err
+}
+
+func readNewBytes(r *bufio.Reader, sz int, n int) ([]byte, int, error) {
+	return readIntoAllocatedBytes(r, sz, n, func(length int) []byte { return make([]byte, length) })
 }
 
 func readArrayLen(r *bufio.Reader, sz int, n *int) (int, error) {