I've seen in the profiles that a big amount of allocations come from creating the slices for `Key` and `Value` fields of `Message` through `readNewBytes`.

I can understand that we don't want to make clients accountable for properly releasing those slices by default, as that can be an unnecessary toil for many. However, in high performance applications it would be nice to have control over this.