@vasily-kirichenko 2024-03-12
It would be great if you added an example (or test) of using a custom key/value allocator. Would it be a `sync.Pool`-based cache or something? When to return a used array back to the pool?