@colega 2024-03-17
Hey, sorry for the delay, I've added a benchmark that shows the usage of these new fields as an example. Would that be enough? I think this is a niche feature that is only relevant in really high-volume environments, so I wouldn't like to distract the newcomers to the landing page with this.

I also fixed the benchmark which was using wrong offset to consume.

Finally the benchmark results:
```
$ benchstat -row .name -col /pooling kafka-bytes-pool
goos: darwin
goarch: arm64
pkg: github.com/segmentio/kafka-go
       │  disabled   │            enabled            │
       │   sec/op    │   sec/op     vs base          │
Reader   4.790µ ± 1%   4.739µ ± 3%  ~ (p=0.239 n=10)

       │   disabled   │            enabled             │
       │     B/s      │     B/s       vs base          │
Reader   407.8Mi ± 1%   412.1Mi ± 3%  ~ (p=0.247 n=10)

       │   disabled   │              enabled               │
       │     B/op     │    B/op     vs base                │
Reader   2089.00 ± 0%   63.50 ± 2%  -96.96% (p=0.000 n=10)

       │  disabled  │            enabled             │
       │ allocs/op  │ allocs/op   vs base            │
Reader   2.000 ± 0%   2.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```